requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.4
djangorestframework==3.17.1
geographiclib==2.1
geopy==2.4.1
idna==3.13
numpy==2.4.4
openrouteservice==2.3.3
pandas==3.0.2
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.2
urllib3==2.6.3
